+++ app/(root)/dashboard/layout.jsx
@@ -1,0 +1,9 @@
+const DashboardLayout = ({ children }) => {
+    return (
+        <main>
+            {children}
+        </main>
+    )
+}
+
+export default DashboardLayout;
