+++ components/shared/Form/loginForm.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const loginForm = () => {
+  return (
+    <div>loginForm</div>
+  )
+}
+
+export default loginForm
+
+
+
