+++ components/shared/NavItems.jsx
@@ -1,14 +1,4 @@
 import Link from "next/link";
-import { Button } from "../ui/button";
-
-import {
-    Tabs,
-    TabsContent,
-    TabsList,
-    TabsTrigger,
-} from "@/components/ui/tabs"
-
-
 
 const NavItems = () => {
