+++ package.json
@@ -26,6 +26,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.49.3",
+    "react-hot-toast": "^2.4.1",
     "react-icons": "^5.0.1",
     "react-redux": "^9.1.0",
     "swiper": "^11.0.5",
